ref: add messages count in preview, add redirect in 404 in case of wrong forum ID, move forum components to its pages
