+++ package.json
@@ -21,6 +21,8 @@
   },
   "devDependencies": {
     "@evilmartians/lefthook": "^1.3.9",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/uuid": "^9.0.6",
     "lerna": "^5.4.3",
     "sass": "^1.69.3"
   },
