+++ packages/client/src/pages/Forum/components/forumPreviewTable/model.ts
