+++ packages/client/src/pages/Forum/components/forumPreviewTable/styles.scss
@@ -1,4 +1,4 @@
-@import '../../App.scss';
+@import '../../../../App.scss';
 
 $tab-count: 2;
