+++ packages/client/src/pages/Forum/components/newTopicForm/style.scss
@@ -1,4 +1,4 @@
-@import '../../App.scss';
+@import '../../../../App.scss';
 
 .newTopicForm {
     display: flex;
