+++ packages/client/src/pages/ForumItemPage/components/addMessageFrame/styles.scss
@@ -1,4 +1,4 @@
-@import '../../App.scss';
+@import '../../../../App.scss';
 
 .addMessageFrame {
     box-sizing: border-box;
