+++ packages/client/src/pages/ForumItemPage/components/forumMessagesList/styles.scss
@@ -1,4 +1,4 @@
-@import '../../App.scss';
+@import '../../../../App.scss';
 
 .forumMessages {
     &__message {
