+++ packages/client/src/pages/index.ts
@@ -3,7 +3,7 @@ export { default as ProfilePage } from './Profile/ProfilePage'
 export { default as Error404 } from './404/Error404'
 export { default as EndPage } from './End/EndPage'
 export { default as AuthPage } from './Auth/AuthPage'
-export { default as ForumPage } from './Forum/forumPage'
+export { default as ForumPage } from './Forum/ForumPage'
 export { default as ForumItemPage } from './ForumItemPage/ForumItemPage'
 export { default as GamePage } from './Game/GamePage'
 export { default as LeaderBoard } from './LeaderBoard/LeaderBoard'
