+++ packages/client/src/utils/getImageUrl.ts
@@ -1,0 +1,5 @@
+const getImageUrl = (name: string | URL): string => {
+  return new URL(name, import.meta.url).href
+}
+
+export default getImageUrl
